Refactor FluentColorPicker to support custom color palettes and enhance HSV color handling
